Renamed camget->reality, working on reality
