+++ interface/reality/machine.go
@@ -1,0 +1,4 @@
+package reality
+
+type Machine struct {
+}
