+++ interface/reality/reality.go
@@ -1,0 +1,10 @@
+package reality
+
+import "github.com/ArchieT/3manchess/game"
+import "github.com/ArchieT/3manchess/player"
+import "log"
+
+type Reality struct {
+	View
+	Machine
+}
